Week 9/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
autogen-agentchat==0.7.5
autogen-core==0.7.5
certifi==2026.2.25
click==8.3.2
cuda-bindings==13.2.0
cuda-pathfinder==1.5.3
cuda-toolkit==13.0.2
faiss-cpu==1.13.2
filelock==3.29.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonref==1.1.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
ollama==0.6.1
opentelemetry-api==1.41.0
packaging==26.1
pillow==12.2.0
protobuf==5.29.6
pyautogen==0.10.0
pydantic==2.13.2
pydantic_core==2.46.2
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.4.1
setuptools==81.0.0
shellingham==1.5.4
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.4
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
zipp==3.23.1
